refactor: extract sibling-coupling detection into analyzer/coupling.rs (#190)
Moves CouplingViolation, build_dir_tree, dir_depth, and compute_dacc out
of analyzer/mod.rs. Also wraps the BFS sweep that walks every parent
directory looking for cross-sibling imports (and the circular
dependencies among them) into a single compute_coupling_violations()
helper. audit() now delegates the entire violation-collection step to
this helper. Behavior is unchanged — bit-identical audit/report output.
